+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.5-wip
 
 * Update video/mp4 mimeType lookup by header bytes.
+* Add image/heic mimeType lookup by header bytes.
+* Add image/heif mimeType lookup by header bytes.
 * Add m4b mimeType lookup by extension.
 * Add `text/markdown` mimeType lookup by extension.
 * Require Dart 3.0.0.